LayoutTests/imported/w3c/web-platform-tests/svg/animations/animate-path-animation-mM-aA-Z-inverse.tentative-expected.txt


FAIL Path animation where coordinate modes of start and end differ (m-M, a-A and Z) assert_equals: expected "M -72.50 32.50 A 157.50 167.50 52.50 1.00 1.00 -7.50 77.50 M 92.50 120.00 A 177.50 187.50 112.50 1.00 1.00 232.50 267.50 Z M 247.50 85.00" but got "m -72.50 32.50 a 157.50 167.50 52.50 1.00 1.00 65.00 45.00 m 100.00 42.50 a 177.50 187.50 112.50 1.00 1.00 140.00 147.50 Z m 90.00 -70.00"
